fix: refactored trainer and entity

+++ src/main/scala/model/entities/World.scala
@@ -6,4 +6,4 @@ object World:
    * @param x coordinate
    * @param y coordinate
    */
-  case class Position(x: Int, y: Int)
+  case class Position(x: Float, y: Float)
